Added "07:05:25.jpg" to "Kanban-photos" branch
In this kanban board photo, I've found a stakeholder and company for me to make a website for, Sudhir Duppati and the Art Committee. Now, I've began asking for his specifications and requirements for the website. Additionally, I've began coding a navigation bar for my website which will have all the webpages on it and appear on every webpage for the user to be easily able to be on the webpage they want to be on. The navigation bar might be subject to change as I receive the specifications and requirements from my stakeholder.