Updated "general.css"
I've updated the general.css file. Some of the updates are changing the a tag which links to content outside the webpage so that it doesn't include an underline and adding a margin to its right. I've also edited the header tag's background colour to black. Additionally, I've made more comments to the file to explain what the code in the file does more extensively.