Added 09:05:24.jpg to the Kanban-photos branch
In this photo of my kanban board, I've finished making a navigation bar and have found the specifications and requirements of my stakeholder. Now, I have started working on creating a footer for my website and after that, plan to continue making the home page of my website by making a background and start making a contacts page by making the inputs and javascript code for it.