Added 10:05:24.jpg to the Kanban-photos branch
In this kanban board photo, I have finished making a footer for my website. Now, I've started making the background of my home page that will also be used for the other web pages. Also, I've split making the contract into two parts: the inputs and the javascript code. This is to decompose the task and make it easier to manage for me. Right now, I'm working on the inputs and plan to work on the javascript code in the future.