Updated style.css
I have updated the style.css file for my website. I have added more tags and classes which ideally improve the look and function of my website. Some key aspects I've added are stylising the footer and adding a class for a background image which will appear on all pages of my website. I have also added the inputs and textarea tags which the user can write in for the contacts page.