Add 13:05:24.jpg to "Kanban-photos" branch
In this kanban photo, I have removed the task to finish the Grok javascript course. This is because I feel it unnecessary to do it as time to finish the website is decreasing and I need to prioritise that. I have also finished making the background for the website and making the inputs for the contact page.