Updated style.css
I have written more css code for the web pages. Specifically, I have added code for the "About" and "Members" pages of my website. I have also removed the css code for the "Contact" page as I have decided to forgo that for simply redirecting the user to their chosen email service with the email for the arts council already filled in.