Deleted "contact.html"
I have decided to delete the "contact.html" file. This is because after speaking to my stakeholder, Sudhir Duppati, he and I decided that a contact page wouldn't necessary and instead I could just have a link to the user's gmail with an email draft that has the council's email address already filled in.