Updated "style-mobile.css"
I have updated the "style-mobile.css" CSS file. Specifically, I have changed the navigation bar to be a dropdown menu of the web page links when the website is on mobile screens. I have also changed the size of elements in the website to be more visible on mobile screens.