Updated "index.html"
I have updated the "index.html" HTML file by changing the navigation so it turns into a dropdown menu when on mobile screens. I have also changed the Javascript code for the image slider so it's its own individual file instead of the file also having code for the dropdown menu.